requirements.txt
kafka-python==1.4.3
Flask==1.0.2
machine
usocket
network
